# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Klar API",
    "description": "German bureaucratic mail \u2192 structured obligations + deadlines.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rag/search": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Search the German bureaucracy knowledge corpus",
        "description": "Returns the top-k vector-similarity matches from the seed corpus (institutions, phrase patterns, deadline conventions). Internal debug surface \u2014 not part of the production frontend's main flow.",
        "operationId": "search_api_rag_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rag/reseed": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Reload the seed corpus into ChromaDB",
        "description": "Upserts every entry from `app/rag/seed.py::SEED_DOCS`.",
        "operationId": "reseed_api_rag_reseed_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetails": {
        "properties": {
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "declared": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Declared"
          },
          "detected": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected"
          }
        },
        "type": "object",
        "title": "ErrorDetails",
        "description": "Optional structured extras attached to an error envelope."
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable identifier \u2014 see docs/06-api-contract.md."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Localized, user-facing copy."
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Every non-2xx response from this API has this shape.\n\nFrontend code should switch on `code` (stable identifier), display\n`message` to the user, and inspect `details` for field-level\ninformation on `VALIDATION_ERROR`s."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RagHit": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "text",
          "score",
          "metadata"
        ],
        "title": "RagHit"
      },
      "RagQuery": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 4
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RagQuery"
      },
      "RagResponse": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/RagHit"
            },
            "type": "array",
            "title": "Hits"
          }
        },
        "type": "object",
        "required": [
          "hits"
        ],
        "title": "RagResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationFieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Dotted path to the offending field."
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ValidationFieldError"
      }
    }
  }
}